+++ src/main/java/frc/robot/RobotContainer.java
@@ -16,8 +16,8 @@ import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.Constants.ControllerConstants;
 import frc.robot.Constants.TunerConstants;
 import frc.robot.Constants.DrivetrainConstants.DriveRequests;
+import frc.robot.commands.StandinCommands;
 import frc.robot.subsystems.PhotonVision;
-import frc.robot.subsystems.StandinCommands;
 import frc.robot.subsystems.Swerve;
 import frc.thunder.LightningContainer;
 import frc.thunder.shuffleboard.LightningShuffleboard;
